requirements.txt
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> dcb09cd (Testing Jack's ability to commit changes)
mysqlclient>=2.0.3,<2.1.0

django-mysql>=3.10.0,<3.11.0

Django>=3.1.4,<3.2.0

djangorestframework>=3.12.2,<3.13.0
<<<<<<< HEAD
<<<<<<< HEAD
=======
boto3
>>>>>>> dcb09cd (Testing Jack's ability to commit changes)
=======

boto3~=1.16.54
>>>>>>> 0d964ae (ECC-420 Validation of metadata in XIS request)

flake8>=3.8.4,<3.9.0

gunicorn>=20.0.4,<20.1.0

martor>=1.5.8,<1.6.0

<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
pandas>=1.1.5,<1.2.0
=======
pandas>=1.1.5,<1.2.0

fsspec>=0.8.5,<0.9.0

numpy>= 1.19.5, < 1.20.0

s3fs
>>>>>>> 0d964ae (ECC-420 Validation of metadata in XIS request)

boto3~=1.16.54

fsspec>=0.8.5,<0.9.0

numpy>= 1.19.5, < 1.20.0

s3fs
=======
=======
>>>>>>> 68c08d4 (django project set up with docker and same api endpoint)
mysqlclient>=2.0.3,<2.1.0

django-mysql>=3.10.0,<3.11.0

Django>=3.1.4,<3.2.0

djangorestframework>=3.12.2,<3.13.0
boto3

flake8>=3.8.4,<3.9.0

gunicorn>=20.0.4,<20.1.0

martor>=1.5.8,<1.6.0








<<<<<<< HEAD
>>>>>>> 86c1339 (django project set up with docker; sample api endpoint)
=======
>>>>>>> 68c08d4 (django project set up with docker and same api endpoint)
=======

=======
pandas>=1.1.5,<1.2.0
>>>>>>> 96c3f05 (added configuration model; added validation helper functions)






>>>>>>> dcb09cd (Testing Jack's ability to commit changes)
